Rename traffic to SEO metrics and move classes to the right level
